requirements.txt
openai
langchain
langchain-openai
langgraph
pymc
xgboost
scikit-learn
scipy
pandas
numpy
requests
beautifulsoup4
psycopg2-binary
redis
python-dotenv
schedule
fastapi
uvicorn
sqlalchemy
aiohttp
lxml
